@@ -121,8 +121,7 @@ public class MixinStringSplitter {
      */
     @Overwrite
     public FormattedText headByWidth(@Nonnull FormattedText text, int width, @Nonnull Style style) {
-        if (text instanceof TextComponent) {
-            TextComponent c = (TextComponent) text;
+        if (text instanceof TextComponent c) {
             if (c.getStyle().getFont().equals(Minecraft.ALT_FONT)) {
                 final float[] maxWidth = {width};
                 final int[] position = {0};
