@@ -20,10 +20,10 @@ package icyllis.modernui.textmc;
 
 import com.google.common.collect.ImmutableList;
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
+import com.mojang.blaze3d.vertex.VertexFormat;
 import icyllis.modernui.ModernUI;
 import net.minecraft.client.renderer.RenderStateShard;
 import net.minecraft.client.renderer.RenderType;
-import org.lwjgl.opengl.GL11;
 
 import javax.annotation.Nonnull;
 import java.util.Objects;
@@ -40,14 +40,10 @@ public class EffectRenderType extends RenderType {
         STATES = ImmutableList.of(
                 RenderStateShard.NO_TEXTURE,
                 RenderStateShard.TRANSLUCENT_TRANSPARENCY,
-                RenderStateShard.NO_DIFFUSE_LIGHTING,
-                RenderStateShard.FLAT_SHADE,
-                RenderStateShard.DEFAULT_ALPHA,
                 RenderStateShard.LEQUAL_DEPTH_TEST,
                 RenderStateShard.CULL,
                 RenderStateShard.LIGHTMAP,
                 RenderStateShard.NO_OVERLAY,
-                RenderStateShard.FOG,
                 RenderStateShard.NO_LAYERING,
                 RenderStateShard.MAIN_TARGET,
                 RenderStateShard.DEFAULT_TEXTURING,
@@ -57,14 +53,10 @@ public class EffectRenderType extends RenderType {
         SEE_THROUGH_STATES = ImmutableList.of(
                 RenderStateShard.NO_TEXTURE,
                 RenderStateShard.TRANSLUCENT_TRANSPARENCY,
-                RenderStateShard.NO_DIFFUSE_LIGHTING,
-                RenderStateShard.FLAT_SHADE,
-                RenderStateShard.DEFAULT_ALPHA,
                 RenderStateShard.NO_DEPTH_TEST,
                 RenderStateShard.CULL,
                 RenderStateShard.LIGHTMAP,
                 RenderStateShard.NO_OVERLAY,
-                RenderStateShard.FOG,
                 RenderStateShard.NO_LAYERING,
                 RenderStateShard.MAIN_TARGET,
                 RenderStateShard.DEFAULT_TEXTURING,
@@ -78,7 +70,7 @@ public class EffectRenderType extends RenderType {
     private EffectRenderType() {
         super(ModernUI.ID + ":text_effect",
                 DefaultVertexFormat.POSITION_COLOR_LIGHTMAP,
-                GL11.GL_QUADS, 256, false, true,
+                VertexFormat.Mode.QUADS, 256, false, true,
                 () -> STATES.forEach(RenderStateShard::setupRenderState),
                 () -> STATES.forEach(RenderStateShard::clearRenderState));
         this.hashCode = Objects.hash(super.hashCode(), STATES);
@@ -87,7 +79,7 @@ public class EffectRenderType extends RenderType {
     private EffectRenderType(String t) {
         super(t,
                 DefaultVertexFormat.POSITION_COLOR_LIGHTMAP,
-                GL11.GL_QUADS, 256, false, true,
+                VertexFormat.Mode.QUADS, 256, false, true,
                 () -> SEE_THROUGH_STATES.forEach(RenderStateShard::setupRenderState),
                 () -> SEE_THROUGH_STATES.forEach(RenderStateShard::clearRenderState));
         this.hashCode = Objects.hash(super.hashCode(), SEE_THROUGH_STATES);
