@@ -19,14 +19,10 @@
 package icyllis.modernui.textmc;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.BufferBuilder;
-import com.mojang.blaze3d.vertex.BufferUploader;
-import com.mojang.blaze3d.vertex.DefaultVertexFormat;
-import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.blaze3d.vertex.*;
 import com.mojang.math.Matrix4f;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.Sheets;
-import org.lwjgl.opengl.GL11;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -135,7 +131,7 @@ public class TextRenderNode {
             g = startG;
             b = startB;
             RenderSystem.disableTexture();
-            builder.begin(GL11.GL_QUADS, DefaultVertexFormat.POSITION_COLOR);
+            builder.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_COLOR);
             for (BaseGlyphRender glyph : mGlyphs) {
                 if ((glyph.mFlags & BaseGlyphRender.COLOR_NO_CHANGE) == 0) {
                     int color = glyph.mFlags;
